+++ KioChess/Engine/Services/Evaluation/EvaluationServiceBase.cs
@@ -34,7 +34,7 @@ public abstract class EvaluationServiceBase : IEvaluationService
     protected byte _doubleRookHorizontalValue;
     protected byte _battaryValue;
     protected byte _openPawnValue;
-    protected sbyte _noPawnsValue;
+    protected short _noPawnsValue;
     private readonly byte _kingShieldPreFaceValue;
     private readonly byte _kingShieldFaceValue;
     private readonly byte _kingZoneOpenFileValue;
@@ -216,7 +216,7 @@ public abstract class EvaluationServiceBase : IEvaluationService
     public byte GetOpenPawnValue() { return _openPawnValue; }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public sbyte GetNoPawnsValue() { return _noPawnsValue; }
+    public short GetNoPawnsValue() { return _noPawnsValue; }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public byte GetPassedPawnValue() { return _passedPawnValue; }
@@ -268,7 +268,7 @@ public abstract class EvaluationServiceBase : IEvaluationService
         _doubleRookHorizontalValue = (byte)(evaluationStatic.DoubleRookHorizontalValue * _unitValue);
         _battaryValue = (byte)(evaluationStatic.BattaryValue * _unitValue);
         _openPawnValue = (byte)(evaluationStatic.OpenPawnValue * _unitValue);
-        _noPawnsValue = (sbyte)(-evaluationStatic.NoPawnsValue * _unitValue);
+        _noPawnsValue = (short)(-evaluationStatic.NoPawnsValue * _unitValue);
 
         _values = new short[12];
         _values[Pieces.WhitePawn] = evaluationProvider.GetPiece(phase).Pawn;
